migrate more user/organization action tests to pytest
