+++ ckanext/unhcr/tests/nose/test_actions.py
@@ -1072,64 +1072,6 @@ class TestAccessRequestUpdate(base.FunctionalTestBase):
         )
 
 
-class TestUpdateSysadmin(base.FunctionalTestBase):
-
-    def test_sysadmin_not_authorized(self):
-        user1 = core_factories.User()
-        user2 = core_factories.User()
-        action = toolkit.get_action("user_update_sysadmin")
-        assert_raises(
-            toolkit.NotAuthorized,
-            action,
-            {"user": user1["name"]},
-            {'id': user1["name"], 'is_sysadmin': True}
-        )
-        assert_raises(
-            toolkit.NotAuthorized,
-            action,
-            {"user": user2["name"]},
-            {'id': user1["name"], 'is_sysadmin': True}
-        )
-
-    def test_sysadmin_invalid_user(self):
-        user = core_factories.Sysadmin()
-        action = toolkit.get_action("user_update_sysadmin")
-        assert_raises(
-            toolkit.ObjectNotFound,
-            action,
-            {"user": user["name"]},
-            {'id': "fred", 'is_sysadmin': True}
-        )
-
-    def test_sysadmin_promote_success(self):
-        admin = core_factories.Sysadmin()
-
-        # create a normal user
-        user = core_factories.User()
-
-        # promote them
-        action = toolkit.get_action("user_update_sysadmin")
-        action({'user': admin['name']}, {'id': user['name'], 'is_sysadmin': True})
-
-        # now they are a sysadmin
-        userobj = model.User.get(user['id'])
-        assert_equals(True, userobj.sysadmin)
-
-    def test_sysadmin_revoke_success(self):
-        admin = core_factories.Sysadmin()
-
-        # create another sysadmin
-        user = core_factories.Sysadmin(fullname='Bob')
-
-        # revoke their status
-        action = toolkit.get_action("user_update_sysadmin")
-        action({'user': admin['name']}, {'id': user['name'], 'is_sysadmin': False})
-
-        # now they are not a sysadmin any more
-        userobj = model.User.get(user['id'])
-        assert_equals(False, userobj.sysadmin)
-
-
 class TestExternalUserUpdateState(base.FunctionalTestBase):
 
     def setup(self):
@@ -1331,128 +1273,6 @@ class TestDatasetCollaboratorCreate(base.FunctionalTestBase):
             }
         )
 
-
-class TestOrganizationMemberCreate(base.FunctionalTestBase):
-
-    def test_internal_user(self):
-        sysadmin = core_factories.Sysadmin(name='sysadmin', id='sysadmin')
-        internal_user = core_factories.User()
-        container = factories.DataContainer()
-
-        toolkit.get_action("organization_member_create")(
-            {'user': sysadmin['name']},
-            {
-                'id': container['id'],
-                'username': internal_user['name'],
-                'role': 'member',
-            }
-        )
-
-        org_list = toolkit.get_action("organization_list_for_user")(
-            {'user': sysadmin['name']},
-            {'id': internal_user['id']}
-        )
-        assert_equals(container['id'], org_list[0]['id'])
-        assert_equals('member', org_list[0]['capacity'])
-
-
-    def test_external_user(self):
-        sysadmin = core_factories.Sysadmin(name='sysadmin', id='sysadmin')
-        external_user = factories.ExternalUser()
-        container = factories.DataContainer()
-
-        action = toolkit.get_action("organization_member_create")
-        assert_raises(
-            toolkit.ValidationError,
-            action,
-            {'user': sysadmin['name']},
-            {
-                'id': container['id'],
-                'username': external_user['name'],
-                'role': 'member',
-            }
-        )
-
-
-class TestUserAutocomplete(base.FunctionalTestBase):
-
-    def test_user_autocomplete(self):
-        sysadmin = core_factories.Sysadmin(name='sysadmin', id='sysadmin')
-        factories.ExternalUser(
-            fullname='Alice External',
-            email='alice@externaluser.com',
-        )
-        core_factories.User(fullname='Bob Internal')
-        core_factories.User(fullname='Carlos Internal')
-        core_factories.User(fullname='David Internal')
-
-        action = toolkit.get_action('user_autocomplete')
-        context = {'user': sysadmin['name']}
-
-        result = action(context, {'q': 'alic'})
-        assert_equals(0, len(result))
-
-        result = action(context, {'q': 'alic', 'include_external': True})
-        assert_equals('Alice External', result[0]['fullname'])
-
-        result = action(context, {'q': 'nal'})
-        fullnames = [r['fullname'] for r in result]
-        assert_in('Bob Internal', fullnames)
-        assert_in('Carlos Internal', fullnames)
-        assert_in('David Internal', fullnames)
-
-        result = action(context, {'q': 'foobar'})
-        assert_equals(0, len(result))
-
-
-class TestUserActions(base.FunctionalTestBase):
-    def test_user_list(self):
-        sysadmin = core_factories.Sysadmin()
-        external_user = factories.ExternalUser()
-        internal_user = core_factories.User()
-        default_user = toolkit.get_action('get_site_user')({ 'ignore_auth': True })
-
-        action = toolkit.get_action('user_list')
-        context = {'user': sysadmin['name']}
-        users = action(context, {})
-        assert_equals(1, len(
-            [
-                u for u in users
-                if u['external']
-                and u['name'] != default_user['name']
-            ])
-        )
-        assert_equals(2, len(
-            [
-                u for u in users
-                if not u['external']
-                and u['name'] != default_user['name']
-            ])
-        )
-
-    def test_user_show(self):
-        sysadmin = core_factories.Sysadmin()
-        external_user = factories.ExternalUser()
-        internal_user = core_factories.User()
-
-        action = toolkit.get_action('user_show')
-        context = {'user': sysadmin['name']}
-        assert_true(action(context, {'id': external_user['id']})['external'])
-        assert_false(action(context, {'id': internal_user['id']})['external'])
-
-    def test_unhcr_plugin_extras_empty(self):
-        user = core_factories.User()
-        context = {'user': user['name']}
-        user = toolkit.get_action('user_show')(context, {'id': user['id']})
-        assert None is user['expiry_date']
-        assert '' == user['focal_point']
-
-    def test_unhcr_plugin_extras_with_data(self):
-        user = factories.ExternalUser(focal_point='Alice')
-        context = {'user': user['name']}
-        user = toolkit.get_action('user_show')(context, {'id': user['id']})
-        assert 'expiry_date' in user
-        assert 'Alice' == user['focal_point']
 
 class TestClamAVActions(base.FunctionalTestBase):
